# api/openapi-lock.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "org-quicko-cliq",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{{base_url}}"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/programs/{program_id}/promoters": {
      "post": {
        "tags": [
          "Promoter"
        ],
        "summary": "Create promoter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "promoter"
                  },
                  "promoter_id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "logo_url": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              },
              "examples": {
                "success": {
                  "value": {
                    "promoter_name": "The Affiliate Man"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully created promoter.\",\r\n    \"data\": {\r\n        \"promoter_id\": \"jwbveie982y938yf93\"\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Promoter"
        ],
        "summary": "Get promoter",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got promoter.\",\r\n    \"data\": {\r\n        \"promoter\": {\r\n            \"promoter_id\": \"jwbveie982y938yf93\",\r\n            \"promoter_name\": \"The Affiliate Man\"\r\n        }\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/programs/{program_id}/promoters/{promoter_id}/members": {
      "post": {
        "tags": [
          "Promoter"
        ],
        "summary": "Invite member",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "member"
                  },
                  "member_id": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "member"
                    ]
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "examples": {
                "success": {
                  "value": {
                    "email": "",
                    "password": "",
                    "first_name": "",
                    "last_name": "",
                    "role": ""
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "promoter_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{promoter_id}}"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully sent an invitation.\",\r\n    \"data\": {\r\n        \"member_id\": \"wne908uj231nc23inc\"\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Promoter"
        ],
        "summary": "Get all members",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "promoter_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{promoter_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got all members.\",\r\n    \"data\": {\r\n        \"members\": [\r\n            {\r\n                \"member_id\": \"wevn1i23yd13hcoiu3bnvkn\",\r\n                \"email\": \"leonardo@mail.com\",\r\n                \"password\": \"2j3n2o30j29j\",\r\n                \"first_name\": \"Leonardo\",\r\n                \"last_name\": \"Da Vinci\"\r\n            },\r\n            {\r\n                \"member_id\": \"ebnf23n89kj-vndkv\",\r\n                \"email\": \"dwight@mail.com\",\r\n                \"password\": \"2j3n2o30j29j\",\r\n                \"first_name\": \"Dwight\",\r\n                \"last_name\": \"Schrute\"\r\n            }\r\n        ]\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/programs/{program_id}/promoters/{promoter_id}/members/{member_id}": {
      "patch": {
        "tags": [
          "Promoter"
        ],
        "summary": "Update role",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "role": ""
                }
              },
              "examples": {
                "success": {
                  "value": {
                    "role": ""
                  }
                },
                "failure- unauthorized": {
                  "value": {
                    "role": ""
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "promoter_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{promoter_id}}"
          },
          {
            "name": "member_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{member_id}}"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully changed user role.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"You are unauthorized to perform this action.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Promoter"
        ],
        "summary": "Remove member",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "promoter_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{promoter_id}}"
          },
          {
            "name": "member_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{member_id}}"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully removed member.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"You are unauthorized to perform this action.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
...
    "/programs/{program_id}/promoters/{promoter_id}/purchases": {
      "get": {
        "tags": [
          "Promoter"
        ],
        "summary": "Get purchases for promoter",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "promoter_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{promoter_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got all purchases referred by promoter.\",\r\n    \"data\": {\r\n        \"purchases\": [\r\n            {\r\n                \"link_id\": \"2m3fj23uicbn\",\r\n                \"email\": \"caesar@mail.com\",\r\n                \"first_name\": \"Julius\",\r\n                \"last_name\": \"Caesar\",\r\n                \"phone\": \"1234567890\",\r\n                \"external_id\": \"qokicmoijv8qj\"\r\n            }\r\n        ]\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
...
  }
}